refactor: Money 클래스 int 타입 long 타입으로 변경

+++ src/main/java/lotto/Money.java
@@ -2,14 +2,14 @@ package lotto;
 
 public class Money {
 
-    private final int money;
+    private final long money;
 
-    public Money(int money) {
+    public Money(long money) {
         validateMoney(money);
         this.money = money;
     }
 
-    private void validateMoney(int money) {
+    private void validateMoney(long money) {
         if (money < 0) {
             throw new IllegalArgumentException("돈이 음수일 수 없습니다.");
         }
